lambda_image_resize/requirements.txt
boto3==1.24.75
botocore==1.27.75
jmespath==1.0.1
Pillow==9.2.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
